blink major: a power-down blinker at 5µA
sleep in power-down mode shows about 5µA on my ampmeter. with a large
resistor for my LED, I'm getting about 300µA when the LED is on.

that'll last very, very long on my large coin battery.

this blinker is trying to make a human-decruyptable blink method that
just displays a counter.